+++ internal/pkg/image/unpacker/squashfs_apptainer.go
@@ -43,10 +43,10 @@ type libBind struct {
 
 // getLibraryBinds returns the library bind mounts required by an elf binary.
 // The binary path must be absolute.
-func getLibraryBinds(binary string) ([]libBind, error) {
+func getLibraryBinds(binary string) (string, []libBind, error) {
 	exe, err := elf.Open(binary)
 	if err != nil {
-		return nil, err
+		return "", nil, err
 	}
 	defer exe.Close()
 
@@ -60,9 +60,9 @@ func getLibraryBinds(binary string) ([]libBind, error) {
 		buf := make([]byte, 4096)
 		n, err := p.ReadAt(buf, 0)
 		if err != nil && err != io.EOF {
-			return nil, err
+			return interp, nil, err
 		} else if n > cap(buf) {
-			return nil, fmt.Errorf("buffer too small to store interpreter")
+			return interp, nil, fmt.Errorf("buffer too small to store interpreter")
 		}
 		// trim null byte to avoid an execution failure with
 		// an invalid argument error
@@ -71,7 +71,7 @@ func getLibraryBinds(binary string) ([]libBind, error) {
 
 	// this is a static binary, nothing to do
 	if interp == "" {
-		return []libBind{}, nil
+		return interp, []libBind{}, nil
 	}
 
 	// run interpreter to list library dependencies for the
@@ -92,10 +92,10 @@ func getLibraryBinds(binary string) ([]libBind, error) {
 	cmd.Env = []string{}
 
 	if err := cmd.Run(); err != nil {
-		return nil, fmt.Errorf("while getting library dependencies: %s\n%s", err, errBuf.String())
+		return interp, nil, fmt.Errorf("while getting library dependencies: %s\n%s", err, errBuf.String())
 	}
-
-	return parseLibraryBinds(buf)
+	bufVal, err := parseLibraryBinds(buf)
+	return interp, bufVal, err
 }
 
 // parseLibrary binds parses `ld-linux-x86-64.so.2 --list <binary>` output.
@@ -215,7 +215,7 @@ func unsquashfsSandboxCmd(unsquashfs string, dest string, filename string, filte
 	}
 
 	// get the library dependencies of unsquashfs
-	libs, err := getLibraryBinds(unsquashfs)
+	interp, libs, err := getLibraryBinds(unsquashfs)
 	if err != nil {
 		return nil, err
 	}
@@ -268,6 +268,9 @@ func unsquashfsSandboxCmd(unsquashfs string, dest string, filename string, filte
 	args = append(args, rootfs)
 
 	// unsquashfs execution arguments
+	if interp != "" {
+		args = append(args, interp)
+	}
 	args = append(args, unsquashfs)
 	args = append(args, opts...)